Started work on redesign
Now the packed_func object is the single object passed around, and is only serialized/deserialized on input and output.

This should result in much easier adaptation to other serialization as well as speed improvements.

At the current time, it does make the API a little clunkier, but this can be improved once all features are supported.